backend/requirements.txt
asgiref==3.7.2
Django==5.0.1
django-rest-framework==0.1.0
djangorestframework==3.14.0
numpy==1.26.3
pandas==2.2.0
python-dateutil==2.8.2
pytz==2023.3.post1
six==1.16.0
sqlparse==0.4.4
tzdata==2023.4
